Rename parser file to flaskparser
