+++ webapp/advantage/views.py
@@ -12,7 +12,7 @@ from webargs.fields import String, Boolean
 # Local
 from webapp.decorators import advantage_checks
 from webapp.login import user_info
-from webapp.advantage.parser import use_kwargs
+from webapp.advantage.flaskparser import use_kwargs
 from webapp.advantage.api import (
     UAContractsAPI,
     CannotCancelLastContractError,
